chore: changes to modules
